@@ -27,18 +27,18 @@
   "author": "Ismael Funes",
   "license": "MIT",
   "devDependencies": {
-    "@angular/common": "5.0.3",
-    "@angular/compiler": "5.0.3",
-    "@angular/compiler-cli": "5.0.3",
-    "@angular/core": "5.0.3",
-    "@angular/forms": "^5.2.7",
-    "@angular/platform-browser": "5.0.3",
-    "@angular/platform-browser-dynamic": "5.0.3",
+    "@angular/common": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@angular/forms": "~11.2.0",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
     "@types/node": "^9.4.6",
     "@ionic/angular": "^5.6.9",
-    "rxjs": "5.5.2",
-    "typescript": "2.4.2",
-    "zone.js": "0.8.18"
+    "typescript": "~4.0.2",
+    "rxjs": "~6.6.0",
+    "zone.js": "~0.10.2"
   },
   "dependencies": {
     "@fortawesome/angular-fontawesome": "^0.1.0",
