@@ -1,5 +1,5 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
-import { IonicModule } from 'ionic-angular';
+import { IonicModule } from '@ionic/angular';
 import { RichTextComponent } from './components/rich-text.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { library } from '@fortawesome/fontawesome-svg-core';
@@ -20,7 +20,7 @@ library.add(fas);
     ]
 })
 export class RichTextModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<RichTextModule> {
         return {
             ngModule: RichTextModule
         };
